Add "useViewStack" and "useRootNode"...
...along with several other changes. Yeah, yeah, this commit should be
smaller, but here's the breakdown:

useNodeStack() has been renamed to useViewStack(). This name better
reflects the intended usage, and along with that the internal variables
were renamed as well. On top of that, useViewStack() returns the current
"top" of the stack (ie stack[0]), instead of the entire stack itself, as
well as a clearStack() callback!

While we are talking about useViewStack(), there were also some
validator functions written to help ensure MainApp is using
useViewStack() correctly, at least during development.

The hook useRootNode() is more of a filler than anything else. It should
be replaced (at least, it's internal definition) with a call to whatever
"nodeStore API" we decide to have once we start using databases. For
now, it simply creates a dummy node on init and uses that for the rest
of the session.

Lastly, inside MainApp itself, this commit implements the hooks
described above. It also moves the display state functions underneath
their corrisponding useState() definition.