Implement MainApp removing top-level nodes
This commit adds a couple key things. First, a function that removes a
node id from the viewNodes array (using push/pop, of course) is defined
as a callback.

Next, since that must be run inside "onNodeDataChange", pretty much
every callback used by MainApp and its hooks had to be wrapped in
useCallback().