Add a common test utils file for all tests
@testing-library recommends that this file be created to provide
overrides/wrappers for default functions that come in the library.

After doing so, a custom render() function was created to include a
wrapper of all contexts created inside the app (which currently is just
the NoteBoardCallbacks context). I also got to learn more about the
"options" that are passed to the library's render() function...