Overhaul test-utils render() and add cleanup()
I noticed that @testing-library/react's render() function created a new
DOM node every time it was invoked. Due to the nature of many components
in this project, this is not okay.

Many tests have been written under
the assumption that the first render() in a test mounts a component, and
subsequent render()s update that same component. Because
@testing-library does not do this, an overhaul was needed. Now, onto the
details of that overhaul...

The render() method exported now strictly uses ReactDOM.render(). It
ensures that everything rendered remains in the same DOM element
throughout testing. It also tracks its calls for later use (see
"cleanup()" below). This method also wraps the given ui with the
 <AllProvidersInApp> component (previously named "AllProviders"), to
ensure that all components tested have access to all contexts that would
be rendered in a normal environment.

The cleanup() method exported removes any DOM elements created by
render(). Because this should be called after every test, warnings are
given if a new test begins and render() was called in the previous test
without calling cleanup() afterward. This is done by tracking calls to
both render() and cleanup(). If, for whatever reason, cleanup() is
intentionally not run, cleanup.ignoreWarnings can be set to true (and
false) to prevent (and restore) logging these warnings.

Lastly, the module's signature is changed, but only slightly. Instead of
exporting all of @testing-library/react, only fireEvent and screen are
exported. This is changed since the primary use of test-utils.jsx is no
longer simply "extending" the @testing-library.