Revert (sort of) test-utils
Okay... So I did a dumb, and realized immediately after all that work I
just did that you could pass a container to the @testing-library's
render() method. So, I decided to reimplement test-utils with a
CustomRenderer class using @testing-library... again...

I also changed/wrote more tests to help with this change. Some of the
tests are a little flaky (they kind of test implementation details), and
they all assume the render/cleanup functions are exported straight from
the CustomRenderer (which I figured is a valid assumption).

Due to these changes, test-utils again matches the signature of
@testing-library.