Implement @testing-library into HookTester
This commit changes HookTester and its associated tests. The main
difference is the render() method is now imported (indirectly) from
@testing-library, which means a DOMRoot is no longer needed for
callHookOn(), and was therefore renamed to callHook().

On top of that, HookTester's prop types were changed slightly, to
reflect which props are optional and which are required.

A couple tests also ensure that overriding console.error() works
correctly. (Originally done to hide PropTypes warnings.)