Import test-utils render method for HookTester
This commit fixes a minor import error; @testing-library was imported
instead of using common/test-utils. Now the wrapped functions are called
and the cleanup checker works!

Or... does it? Actually, this commit fixes a minor bug where the cleanup
checker warnings did not output the file path of the offending test!
(This was due to an incorrect variable name and works correctly now.)